Add a commands to delete user
Hey Chad,
Suhel here. I would like to tell you that there is no command in this readme.md file for deleting a user after if needs to be deleted e.g : we forget the password for that user etc.
I have added the command for deleting a user.
I hope you will merge this PR.
Thank you.